backend/requirements.txt
vanna==0.7.0
flask==3.0.0
flask-cors==4.0.0
numpy<2.0.0,>=1.26.0
faiss-cpu==1.8.0
python-dotenv==1.0.0
openai>=1.0.0
